Extract image service from CRI
